refactored files structure
