+++ cishouseholds/pipeline/high_level_transformations.py
@@ -87,14 +87,14 @@ from cishouseholds.impute import impute_latest_date_flag
 from cishouseholds.impute import impute_outside_uk_columns
 from cishouseholds.impute import impute_visit_datetime
 from cishouseholds.impute import merge_previous_imputed_values
-from cishouseholds.mapping import column_name_maps
+from cishouseholds.pipeline.mapping import column_name_maps
+from cishouseholds.pipeline.regex_patterns import at_school_pattern
+from cishouseholds.pipeline.regex_patterns import at_university_pattern
+from cishouseholds.pipeline.regex_patterns import retired_regex_pattern
+from cishouseholds.pipeline.regex_patterns import self_employed_regex
+from cishouseholds.pipeline.regex_patterns import work_from_home_pattern
 from cishouseholds.pipeline.timestamp_map import cis_digital_datetime_map
 from cishouseholds.pyspark_utils import get_or_create_spark_session
-from cishouseholds.regex_patterns import at_school_pattern
-from cishouseholds.regex_patterns import at_university_pattern
-from cishouseholds.regex_patterns import retired_regex_pattern
-from cishouseholds.regex_patterns import self_employed_regex
-from cishouseholds.regex_patterns import work_from_home_pattern
 from cishouseholds.validate_class import SparkValidate
